Created and integrated session data package
