+++ backend/modules/track_handler.py
@@ -20,6 +20,7 @@ if TYPE_CHECKING:
     from modules.connection import Connection
     from modules.filter_api_interface import FilterAPIInterface
 
+
 class TrackHandler(MediaStreamTrack):
     """Handles and distributes an incoming audio track to multiple subscribers."""
