+++ backend/session/data/session/session_dict.py
@@ -1,0 +1,66 @@
+"""Provide the `SessionDict` TypedDict.
+
+Use for type hints and static type checking without any overhead during runtime.
+"""
+from typing import TypedDict, Any
+
+from custom_types.note import NoteDict
+from session.data.participant import ParticipantDict
+
+
+class SessionDict(TypedDict):
+    """TypedDict for session configuration data.
+
+    Attributes
+    ----------
+    id : str, default ""
+        Unique session id generated by the backend after saving the session.  When
+        creating a new session, this field is initially set to an empty string by the
+        client. Read only for client.
+    title : str
+        Experiment title.
+    description : str
+        Experiment description.
+    date : int
+        Planned starting date / time of the experiment in milliseconds since January 1,
+        1970, 00:00:00 (UTC).
+    time_limit : int
+        Experiment time limit in milliseconds since January 1, 1970, 00:00:00 (UTC).
+    record : bool
+        Whether the experiment is to be recorded.
+    participants : list of custom_types.participant.ParticipantDict
+        List of invited participants.
+    creation_time : int, default 0
+        Time the experiment was created, if an experiment still exists.  A value grater
+        than 0 indicates that there is a running experiment for this session.  Time in
+        milliseconds since January 1, 1970, 00:00:00 (UTC).  Read only for client.
+    start_time : int, default 0
+        Time the experiment started in milliseconds since January 1, 1970, 00:00:00
+        (UTC).  Read only for client.
+    end_time : int, default 0
+        Time the experiment ended in milliseconds since January 1, 1970, 00:00:00 (UTC).
+        Read only for client.
+    notes : list of custom_types.note.NoteDict
+        List of `NoteDict`s the experimenter added during the experiment.  Read only for
+        client.
+    log : Any, optional
+        Log created by the backend.  Read only for client.
+
+    See Also
+    --------
+    Data Types Wiki :
+        https://github.com/TUMFARSynchorny/experimental-hub/wiki/Data-Types#session
+    """
+
+    id: str
+    title: str
+    description: str
+    date: int
+    time_limit: int
+    record: bool
+    participants: list[ParticipantDict]
+    creation_time: int
+    start_time: int
+    end_time: int
+    notes: list[NoteDict]
+    log: Any
